fix offers discount field to float, add offers admin view

+++ django_website/products/models.py
@@ -13,7 +13,7 @@ class Product(models.Model):
 class Offer(models.Model):
   code = models.CharField(max_length=10)
   description = models.CharField(max_length=255)
-  discount = models.IntegerField()
+  discount = models.FloatField()
   
 
 #
